ロジック部分をhook化
